requirements.txt
# Flask for API
Flask==3.0.0
Flask-Cors==4.0.0

# File handling and downloads
requests==2.31.0

# PDF processing
pdfplumber==0.11.0
PyPDF2==3.0.1

# DOCX processing
python-docx==1.1.0

# Image processing and OCR
Pillow==10.1.0
pytesseract==0.3.13
opencv-python==4.9.0.80
numpy==1.26.2

# NLP and Machine Learning
transformers==4.36.2
scikit-learn==1.3.2
joblib==1.3.2

# Note: The following model will be downloaded automatically on first use:
# - facebook/bart-large-cnn (~1.6GB) - English summarization only
# 
# Total model storage: ~1.6GB in ~/.cache/huggingface/
#
# Supported Languages:
# - English: Direct processing and summarization